+++ src/test/java/edu/hm/hafner/analysis/parser/PmdParserTest.java
@@ -6,9 +6,10 @@ import edu.hm.hafner.analysis.AbstractParserTest;
 import edu.hm.hafner.analysis.Issue;
 import edu.hm.hafner.analysis.Report;
 import edu.hm.hafner.analysis.Severity;
-import static edu.hm.hafner.analysis.assertj.Assertions.*;
 import edu.hm.hafner.analysis.assertj.SoftAssertions;
 import edu.hm.hafner.analysis.parser.pmd.PmdParser;
+
+import static edu.hm.hafner.analysis.assertj.Assertions.*;
 
 /**
  * Tests the extraction of PMD analysis results.
@@ -44,7 +45,8 @@ class PmdParserTest extends AbstractParserTest {
                 .hasLineStart(54)
                 .hasLineEnd(61)
                 .hasPackageName("com.avaloq.adt.env.internal.ui.actions")
-                .hasFileName("C:/Build/Results/jobs/ADT-Base/workspace/com.avaloq.adt.ui/src/main/java/com/avaloq/adt/env/internal/ui/actions/CopyToClipboard.java");
+                .hasFileName(
+                        "C:/Build/Results/jobs/ADT-Base/workspace/com.avaloq.adt.ui/src/main/java/com/avaloq/adt/env/internal/ui/actions/CopyToClipboard.java");
     }
 
     @Test
@@ -62,6 +64,58 @@ class PmdParserTest extends AbstractParserTest {
                 .hasCategory("Code Size")
                 .hasSeverity(Severity.WARNING_NORMAL)
                 .hasMessage("The method 'parse' has a Cyclomatic Complexity of 10.");
+    }
+
+    /**
+     * Parses a warning log with errors.
+     *
+     * @see <a href="http://issues.jenkins-ci.org/browse/JENKINS-54736">Issue 54736</a>
+     */
+    @Test
+    void issue54736() {
+        Report report = parseInPmdFolder("issue54736.xml");
+
+        assertThat(report).hasSize(4 + 21).hasSeverities(21, 0, 4, 0);
+        assertThat(report.get(4)).hasSeverity(Severity.ERROR)
+                .hasFileName(
+                        "/Users/jordillach/DemoTenants/Tenants/vhosts/pre.elperiodico.com/themes/default/articleTemplates/forceOpinion.s.jsp")
+                .hasMessage(
+                        "Error while parsing /Users/jordillach/DemoTenants/Tenants/vhosts/pre.elperiodico.com/themes/default/articleTemplates/forceOpinion.s.jsp")
+                .hasDescription(
+                        "\"<!--\" ...net.sourceforge.pmd.PMDException: Error while parsing /Users/jordillach/DemoTenants/Tenants/vhosts/pre.elperiodico.com/themes/default/articleTemplates/forceOpinion.s.jsp\n"
+                                + "\tat net.sourceforge.pmd.SourceCodeProcessor.processSourceCode(SourceCodeProcessor.java:99)\n"
+                                + "\tat net.sourceforge.pmd.SourceCodeProcessor.processSourceCode(SourceCodeProcessor.java:51)\n"
+                                + "\tat net.sourceforge.pmd.processor.PmdRunnable.call(PmdRunnable.java:78)\n"
+                                + "\tat net.sourceforge.pmd.processor.PmdRunnable.call(PmdRunnable.java:24)\n"
+                                + "\tat java.util.concurrent.FutureTask.run(FutureTask.java:266)\n"
+                                + "\tat java.util.concurrent.Executors$RunnableAdapter.call(Executors.java:511)\n"
+                                + "\tat java.util.concurrent.FutureTask.run(FutureTask.java:266)\n"
+                                + "\tat java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1149)\n"
+                                + "\tat java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:624)\n"
+                                + "\tat java.lang.Thread.run(Thread.java:748)\n"
+                                + "Caused by: net.sourceforge.pmd.lang.jsp.ast.ParseException: Encountered \" \"</\" \"</ \"\" at line 22, column 1.\n"
+                                + "Was expecting one of:\n"
+                                + "    <EOF>\n"
+                                + "    \"<\" ...\n"
+                                + "    \"<![CDATA[\" ...\n"
+                                + "    \"<%--\" ...\n"
+                                + "    \"<%!\" ...\n"
+                                + "    \"<%=\" ...\n"
+                                + "    \"<%\" ...\n"
+                                + "    \"<%@\" ...\n"
+                                + "    \"<script\" ...\n"
+                                + "    <EL_EXPRESSION> ...\n"
+                                + "    <UNPARSED_TEXT> ...\n"
+                                + "    \n"
+                                + "\tat net.sourceforge.pmd.lang.jsp.ast.JspParser.generateParseException(JspParser.java:1846)\n"
+                                + "\tat net.sourceforge.pmd.lang.jsp.ast.JspParser.jj_consume_token(JspParser.java:1725)\n"
+                                + "\tat net.sourceforge.pmd.lang.jsp.ast.JspParser.CompilationUnit(JspParser.java:55)\n"
+                                + "\tat net.sourceforge.pmd.lang.jsp.JspParser.parse(JspParser.java:41)\n"
+                                + "\tat net.sourceforge.pmd.SourceCodeProcessor.parse(SourceCodeProcessor.java:111)\n"
+                                + "\tat net.sourceforge.pmd.SourceCodeProcessor.processSource(SourceCodeProcessor.java:175)\n"
+                                + "\tat net.sourceforge.pmd.SourceCodeProcessor.processSourceCode(SourceCodeProcessor.java:96)\n"
+                                + "\t... 9 more");
+
     }
 
     /**
@@ -117,7 +171,8 @@ class PmdParserTest extends AbstractParserTest {
 
         int expectedSize = 4;
         assertThat(report).hasSize(expectedSize);
-        assertThat(report.filter(Issue.byPackageName("com.avaloq.adt.env.core.db.plsqlCompletion"))).hasSize(expectedSize);
+        assertThat(report.filter(Issue.byPackageName("com.avaloq.adt.env.core.db.plsqlCompletion"))).hasSize(
+                expectedSize);
         assertThat(report).hasSeverities(0, 0, 4, 0);
     }